+++ src/main/java/com/yuuko/core/scheduler/tasks/TruncateDatabaseTask.java
@@ -1,15 +1,12 @@
 package com.yuuko.core.scheduler.tasks;
 
-import com.yuuko.core.database.DatabaseFunctions;
 import com.yuuko.core.metrics.handlers.MetricsManager;
 import com.yuuko.core.scheduler.Task;
 
-public class UploadMetricsTask implements Task {
+public class TruncateDatabaseTask implements Task {
 
     @Override
     public void handle() {
-        DatabaseFunctions.updateMetricsDatabase();
-        MetricsManager.reset();
+        MetricsManager.truncate();
     }
-
 }